requirements.txt
networkx>=3.0,<4
pydantic>=2.0,<3
pytest>=7.0,<9
PyYAML>=6.0,<7
